:sparkles: Feat: 프로필 이미지 기본 편집 상태 추가
